% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bg_window_summary.R
\name{bg_window_summary}
\alias{bg_window_summary}
\title{bg_window_summary}
\usage{
bg_window_summary(cases)
}
\arguments{
\item{cases}{data.frame that includes column names "cp_id", "date_surgery", "surgery_end_dts"}
}
\value{
Returns a data frame with cp_id, start,end (of the BG Window) min, max, count_values,
& count_180.
}
\description{
wrapper function for bg_window,
}
\details{
takes a data frame with columns cp_id, date_surgery, and surgery_end_dts
and returns the bg_window data for each row
}
\keyword{bg}

