% Generated by roxygen2 (4.0.1): do not edit by hand
\name{connection_string}
\alias{connection_string}
\title{connection_string}
\usage{
connection_string(resource, database = NULL, uid = getOption("Apollo_uid"),
  pwd = getOption("Apollo_pwd"))
}
\arguments{
  \item{resource}{- Alias for the database connection}
}
\description{
queries EDW using windows credentials
}
\keyword{sql}

