% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cusumPlot.R
\name{cusumPlot}
\alias{cusumPlot}
\title{cusumPlot}
\usage{
cusumPlot(O, E, sort.var, ...)
}
\arguments{
\item{O}{vector of observed rate/probability}

\item{E}{vector of expected rate /probability}

\item{sort.var}{variable to sort to CUSUM}

\item{...,}{additional parameters passed to base plot & lines functions}
}
\description{
displays a straight-forward CUSUM plot
}
\examples{
O <- c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0)
E <- c(0.1, 0.1, 0.1, 0.1, 0.2, 0.2, 0.3, 0.3, 0.4, 0.4, 0.1, 0.1,
       0.1, 0.1, 0.2, 0.2, 0.3, 0.3, 0.4, 0.4)
sort.var <- 1:20
cusumPlot(O, E, sort.var, xlab = "case #", ylab = "avoided cases", main = "Title", col = "dark blue")
}
\keyword{cusum}

