% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dotplot.errors}
\alias{dotplot.errors}
\title{dotplot.errors}
\usage{
dotplot.errors(x, myTheme = simpleTheme(pch = 19, col = 1),
  qlabel = "Estimated rate", add.text.to.qlabel = "", type.bar = "CI",
  conf.level = 0.95, end.length = 0.05, reorder.groups = TRUE,
  reordering = "decreasing", label.define = list(), reference.line = NULL,
  bar.color = 1, horizontal = TRUE, ...)
}
\arguments{
  \item{title}{The title of the measure}

  \item{group}{vector of the names of each observation (ie.
  hospital, provider)}

  \item{measure}{vector of measure values}

  \item{upper}{vector with each upper confidence interval}

  \item{lower}{vector with each lower confidence interval}

  \item{benchmark}{the average or benchmark value}
}
\value{
plots a simple dotplot with title and benchmark as a
horizontal dashed line.
}
\description{
displays a basic dotplot with confidence bars
}
\keyword{dotplot}

