% Generated by roxygen2 (4.0.0): do not edit by hand
\name{edw_tables}
\alias{edw_tables}
\title{edw_tables}
\usage{
edw_tables(catalog, schema, tableName, resource = "Phloston")
}
\arguments{
\item{schema}{- name of the desired schema}

\item{tableName}{- name of the desired table}

\item{resource}{- Alias for the database connection}
}
\description{
queries EDW using windows credentials
}
\examples{
edw_tables("Phloston")
}
\keyword{sql}

