% Generated by roxygen2 (4.0.0): do not edit by hand
\name{edwFetch}
\alias{edwFetch}
\title{edwFetch}
\usage{
edwFetch(resource = "Phloston", schemaName = NULL, tableName = NULL,
  colnames = FALSE, max = 10)
}
\arguments{
\item{resource}{- Alias for the database connection}

\item{schemaName}{- desired schema}

\item{tableName}{- desired table}

\item{colnames}{- desired columns}

\item{max}{= TOP n rows from the query results}
}
\description{
read some or all of a table from the EDW into a data frame
}
\examples{
edw_tables("Phloston")
}
\keyword{sql}

