% Generated by roxygen2 (4.0.0): do not edit by hand
\name{edwQuery}
\alias{edwQuery}
\title{edw_query}
\usage{
edwQuery(schema, table_name, top = 10, fields = "*", where = "1=1",
  resource = "Phloston")

edwQuery(schema, table_name, top = 10, fields = "*", where = "1=1",
  resource = "Phloston")
}
\arguments{
\item{schema}{- name of the desired schema}

\item{table_name}{- name of the desired table}

\item{top}{- integer - use a large number for all records}

\item{fields}{- single string of field names separated by commas}

\item{where}{- additional where clauses}

\item{resource}{- Alias for the database connection}

\item{schema}{- name of the desired schema}

\item{table_name}{- name of the desired table}

\item{top}{- integer - use a large number for all records}

\item{fields}{- single string of field names separated by commas}

\item{where}{- additional where clauses}

\item{resource}{- Alias for the database connection}
}
\description{
queries EDW using windows credentials

queries EDW using windows credentials
}
\examples{
sql_query(")
sql_query(")
}
\keyword{sql}

