% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edwSave.R
\name{edwSave}
\alias{edwSave}
\title{edwSave}
\usage{
edwSave(df = mtcars, tablename, resource = "Work", safer = TRUE,
  DSN = FALSE, uid = NULL, pwd = NULL, message = FALSE, ...)
}
\arguments{
\item{df}{- data.frame}

\item{tablename}{- name of table on database}

\item{resource}{- Alias for the database connection

\describe{
  \item{custom}{allows for a connection to any database server. The connection string must be passed using the custom parameter.}
  \item{Work}{ProvidenceEpic_Work on EDWDBDev}
}}

\item{safer}{logical. If true, create a non-existing table but only allow appends to an existing table. 
If false, allow sqlSave to attempt to delete all the rows of an existing table, or to drop it}

\item{DSN}{if TRUE, use a DSN as specified in `resource` parameter instead of a connection string (ie. for Axis PATS: resource = "TSI_32", DSN = TRUE)}

\item{uid}{username for connections if necessary. Default is NULL, which will then use the user's system credentials.}

\item{pwd}{password for connections that don't use the user's system credentials. Default is NULL.}

\item{message}{display the time required to complete?}

\item{...}{arguments to be passed to RODBC::odbcDriverConnect}
}
\value{
\item{status_message}{}
}
\description{
Saves a data frame to the specified resource
#'
}
\keyword{sql}

