% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edwSQL.R
\name{edwSQL}
\alias{edwSQL}
\title{edwSQL}
\usage{
edwSQL(sql = "SELECT TOP 10 * FROM Event_Cath", resource = "Apollo",
  file = TRUE, ...)
}
\arguments{
\item{sql}{- sql query or the path & name of a .sql file}

\item{resource}{- Alias for the database connection

\describe{
  \item{Trantor}{EDWDBProd}
  \item{Phloston}{EDWDBDev}
  \item{ClinicalAnalytics}{\\\\wn2591\\PremierPRD}
  \item{Apollo}{\\\\wn1444 Apollo_Spokane}
  \item{Eva}{\\\\wn1444 Testing}
}}

\item{file}{switch indicating if sql parameter is a .sql file or a command. Default is TRUE (requires .sql file)}
}
\value{
\item{data}{Query results as a data frame}
\item{fields}{field names}
\item{elapsed_seconds}{elapsed system time to run the query}
\item{status_message}{the number of records & variables in the results}
}
\description{
passes a SQL query to the EDW. Commented lines in the .sql file must be preceded by '--'. Blocked comments are not supported.
}
\keyword{sql}

