% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edwTables.R
\name{edwTables}
\alias{edwTables}
\title{edwTables}
\usage{
edwTables(catalog = NULL, schema = NULL, tableName = NULL,
  tableType = NULL, resource = "Phloston")
}
\arguments{
\item{catalog}{- name of the desired database}

\item{schema}{- name of the desired schema}

\item{tableName}{- name of the desired table}

\item{resource}{- Alias for the database connection}
}
\description{
queries EDW using windows credentials
}
\keyword{sql}

