% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CI.R
\name{get_CI}
\alias{get_CI}
\title{get_CI}
\usage{
get_CI(name, num, den)
}
\arguments{
\item{name}{Vector of names (unit, group, physician, etc.)}

\item{num}{Vector of numerators}

\item{den}{vector of denominators}
}
\description{
Determine the confidence intervals for each group
based on numerators/denominators
}
\author{
Rollie Parrish
}
