% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extended_property.R
\name{get_extended_property}
\alias{get_extended_property}
\title{get_extended_property}
\usage{
get_extended_property(conn, db, schema, table, property)
}
\arguments{
\item{conn}{}

\item{db}{}

\item{schema}{}

\item{table}{}

\item{property}{}
}
\description{
Gets the extended properties in SQL server for
the specified object (database, schema, table)
deferring column properties to a future separate function
}
