% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/medd.R
\name{medd}
\alias{medd}
\title{medd}
\usage{
medd(hydrocodone_po = 0, morphine_iv = 0, morphine_po = 0,
  hydromorphone_iv = 0, fentanyl_iv = 0, fentanyl_td = 0,
  hydromorphone_po = 0, methadone_iv = 0, oxycodone_po = 0,
  oxymorphone_po = 0, sufentanil_iv = 0, nalbuphine_iv = 0,
  tramadol_iv = 0)
}
\arguments{
\item{hydrocodone_po}{PO hydrocodone  total}

\item{morphine_iv}{IV morphine  total}

\item{hydromorphone_iv}{IV hydromorphone total}

\item{fentanyl_iv}{IV fentaNYL total}

\item{fentanyl_td}{transdermal fentaNYL total}

\item{hydromorphone_po}{PO hydromorphone total}

\item{methadone_iv}{methadone  total}

\item{oxycodone_po}{PO Oxycodone total}

\item{oxymorphone_po}{PO Oxymorphone total}

\item{sufentanil_iv}{IV sufentanil}

\item{tramadol_iv}{IV Tramadol}
}
\value{
a list of the total MEDD values
}
\description{
converts the 24hr total dosages of various opioid medications and routes
into a Morphine Equivalent Daily Dose (MEDD)
}
\examples{
data <- data.frame(names=c("john", "jane", "peter"),
                 hydrocod_po=c(5,10,NA),
                 morphine_iv=c(0,10,1),
                 hydromorphone_iv = c(1,2,3),
                 fentanyl_iv = c(2,4,6),
                 fentanyl_td = c(1,1,1),
                 hydromorphone_po = c(NA, NA, 2),
                 methadone_iv = c(1,2,3),
                 oxycodone_po = c(NA, NA, 1),
                 oxymorphone_po = c(2,4,6),
                 sufentanil_iv = c(2,3,4),
                 tramadol_iv = c(1,NA,2)
                  )
data[is.na(data)] <- 0
data$total <- mapply(medd,
     hydrocodone_po = data$hydrocod_po,
     morphine_iv = data$morphine_iv,
     hydromorphone_iv = data$hydromorphone_iv,
     fentanyl_iv = data$fentanyl_iv,
     fentanyl_td = data$fentanyl_td,
     hydromorphone_po = data$hydromorphone_po,
     methadone_iv = data$methadone_iv,
     oxycodone_po = data$oxycodone_po,
     oxymorphone_po = data$oxymorphone_po,
     sufentanil_iv = data$sufentanil_iv,
     tramadol_iv = data$tramadol_iv
     )

as.data.frame(t(data))
}
\keyword{medd}

