% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_template_VA.R
\name{project_template_VA}
\alias{project_template_VA}
\title{create a project_template}
\usage{
project_template_VA(project_path, ...)
}
\arguments{
\item{project_path}{Folder where project will be located}

\item{...}{parameters passed through to other functions}
}
\description{
create a project_template
}
