% Generated by roxygen2 (4.0.1): do not edit by hand
\name{redcap_write_df}
\alias{redcap_write_df}
\title{redcap_write_df}
\usage{
redcap_write_df(ds, redcap_uri, token, verbose = TRUE, cert_location = NULL)
}
\arguments{
  \item{ds}{dataframe to be written}

  \item{redcap_uri}{URL to the REDCap API folder}

  \item{token}{REDCap token (must have Import Data
  enabled)}

  \item{verbose}{(default = TRUE)}

  \item{token}{}

  \item{cert_location}{}
}
\value{
??
}
\description{
writes a dataframe to a REDCap project
}
\examples{
mydata <- data.frame(study_id = c(1,2,3),
                    name = c("apple", "banana", "pear")
                    )

uri <- "https://ampa.org/redcap/api/"
token <- "088E8B8639976E90837B34769C4978E5"

test <- redcap_write_df(mydata, uri, token)
}
\keyword{redcap}

