% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email.R
\name{send.email}
\alias{send.email}
\title{send_email.R}
\usage{
send.email(body = NULL, subject = NULL,
  to = "rollie.parrish@providence.org",
  from = "rollie.parrish@providence.org", sig = TRUE, ...)
}
\arguments{
\item{body}{message body in plain text or markdown. Will be converted to HTML}

\item{subject}{email subject}

\item{to}{list of recipients}

\item{from}{default is <rollie.parrish@providence.org>}

\item{sig}{append Rollie's signature block? default = TRUE}

\item{...}{additional command}
}
\description{
sends an HTML formatted email. Supports basic markdown formatting
and adds a default signature
}

