% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_extended_property.R
\name{set_extended_property}
\alias{set_extended_property}
\title{set_extended_property}
\usage{
set_extended_property(conn, db, schema, table, property, value)
}
\arguments{
\item{conn}{}

\item{schema}{}

\item{table}{}

\item{value}{}

\item{catalog}{}

\item{name}{}
}
\description{
Sets the extended property in SQL server for
the specified object (database, schema, table)
deferring column properties to a future separate function
}
