% Generated by roxygen2 (4.0.0): do not edit by hand
\name{sql_query}
\alias{sql_query}
\title{apollo_query}
\usage{
sql_query(sql, dsn = "", file = TRUE)

sql_query(sql, dsn = "", file = TRUE)
}
\arguments{
  \item{sql}{}

  \item{uid}{userID}

  \item{pwd}{password}

  \item{file}{(defaults to TRUE)}

  \item{sql}{}

  \item{dsn}{}

  \item{file}{(defaults to TRUE)}
}
\description{
queries Apollo Server as configured in system DSN

queries SQL Server as configured in system DSN
}
\examples{
sql_query("select * from table")
sql_query("select * from table")
}
\keyword{sql}

