% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssi.R
\docType{data}
\name{ssi}
\alias{ssi}
\title{SSI.}
\format{A data frame with 5975 rows and 5 variables:
\describe{
  \item{surgery_date}{date of surgery}
  \item{procCode}{NHSN Procedure Code: COLO = Colon surgery, HPRO = Hip Prosthesis, 
  KPRO = Knee Prosthesis, HYST = Abdominal Hysterectomy}
  \item{surgeon_name}{Fake Surgeon name (from `randomNames` package)}
  \item{observed}{Observed SSI. 1 = Yes, 0 = No}
  \item{expected}{Risk of SSI.}
  ...
}}
\usage{
ssi
}
\description{
A fake dataset containing the observed and expected 
risk of SSI.
}
\keyword{datasets}

