% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apollo_query.R
\name{apollo_query}
\alias{apollo_query}
\title{apollo_query}
\usage{
apollo_query(sql, uid, pwd, file = TRUE)
}
\arguments{
\item{sql}{}

\item{uid}{userID}

\item{pwd}{password}

\item{file}{(defaults to TRUE)}
}
\description{
queries Apollo Server as configured in system DSN
}
\keyword{sql}

