% Generated by roxygen2 (4.0.0): do not edit by hand
\name{apollo_query}
\alias{apollo_query}
\title{apollo_query}
\usage{
apollo_query(sql, uid = "roapollo", pwd, file = TRUE)
}
\arguments{
  \item{sql}{}

  \item{uid}{userID}

  \item{pwd}{password}

  \item{file}{(defaults to TRUE)}
}
\description{
queries Apollo Server as configured in system DSN
}
\examples{
sql_query("select * from table")
}
\keyword{sql}

