% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bg_window.R
\name{bg_window}
\alias{bg_window}
\title{bg_window}
\usage{
bg_window(cp_id, surgery_date, surgery_end_time)
}
\arguments{
\item{cp_id}{CommunityPatientID}

\item{surgery_date}{The date of surgery in ISO format}

\item{surgery_end_time}{The surgery end time in HH:MM}
}
\value{
Returns a data frame with cp_id, start,end (of the BG Window) min, max, count_values,
& count_180.
}
\description{
Get the BG data from the 18-24 hour post-op window,
}
\details{
returns the start & end time of the BG window, the minimum & maximum
BG values, the total number of values and the number over 180.
}
\examples{
## single
cp_id <- c("1501668269")
surgery_date <- "2014-10-02"
surgery_end_time <- "13:22"

# bg_window(cp_id, surgery_date, surgery_end_time)
}
\keyword{bg}

