% Generated by roxygen2 (4.0.0): do not edit by hand
\name{connection_string}
\alias{connection_string}
\title{connection_string}
\usage{
connection_string(resource, database = NULL, uid, pwd)
}
\arguments{
  \item{resource}{- Alias for the database connection}
}
\description{
queries EDW using windows credentials
}
\examples{
connection_string("Trantor")
connection_string("Aurora")
}
\keyword{sql}

