% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connection_string.R
\name{connection_string}
\alias{connection_string}
\title{connection_string}
\usage{
connection_string(resource = "custom", custom,
  uid = getOption("Apollo_uid"), pwd = getOption("Apollo_pwd"))
}
\arguments{
\item{resource}{- Alias for the database connection
\describe{
  \item{custom}{custom database connection string, as specified in the custom parameter}
  \item{EDWDBProd}{EDWDBProd}
  \item{Work}{ProvideneEpic_Work}
  \item{Trantor}{EDWDBProd}
  \item{Phloston}{EDWDBDev}
  \item{PHC_CA}{PHC_CA}
  \item{ClinicalAnalytics}{\\\\wn2591\\PremierPRD}
  \item{Apollo}{\\\\wn1444 Apollo_Spokane}
}}

\item{custom}{a valid SQL connection string. used to connect to other resources (ie. a local SQL Express database)}

\item{uid}{userid for those resources that require one}

\item{pwd}{password for those resources that require one. Do not include verbatim passwords in scripts.}
}
\value{
Returns a connection string for RODBC
}
\description{
queries EDW using windows credentials
}
\keyword{sql}

