% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_string.R
\name{connection_string}
\alias{connection_string}
\title{connection_string}
\usage{
connection_string(resource = "custom", custom, uid = NULL, pwd = NULL)
}
\arguments{
\item{resource}{- Alias for the database connection
\describe{
  \item{custom}{custom database connection string, as specified in the custom parameter}
  \item{Trantor}{server_1_name}
  \item{Phloston}{server_2_name}
}}

\item{custom}{a valid SQL connection string. used to connect to other resources (ie. a local DuckDB database)}

\item{uid}{userid for those resources that require one}

\item{pwd}{password for those resources that require one. Never include verbatim passwords in scripts.}
}
\value{
Returns a connection string for odbc()
}
\description{
Used to query a SQL Server instance using windows credentials
}
\keyword{sql}
