% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connection_string.R
\name{connection_string}
\alias{connection_string}
\title{connection_string}
\usage{
connection_string(resource, database = NULL, uid = getOption("Apollo_uid"),
  pwd = getOption("Apollo_pwd"))
}
\arguments{
\item{resource}{- Alias for the database connection
\describe{
  \item{EDWDBProd}{EDWDBProd}
  \item{Trantor}{EDWDBProd}
  \item{Phloston}{EDWDBDev}
  \item{ClinicalAnalytics}{\\\\wn2591\\PremierPRD}
  \item{Apollo}{\\\\wn1444 Apollo_Spokane}
  \item{Eva}{\\\\wn1444 Testing}
}}
}
\value{
Returns a connection string for RODBC
}
\description{
queries EDW using windows credentials
}
\keyword{sql}

