% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cusumPlot}
\alias{cusumPlot}
\title{cusumPlot}
\usage{
cusumPlot(O, E, sort.var, Col = "black")
}
\arguments{
  \item{O}{vector of observed rate/probability}

  \item{E}{vector of expected rate /probability}

  \item{eVar}{estimated variance of event (calculated if
  NULL)}

  \item{sor.var}{variable to sort to CUSUM}

  \item{Col}{color for the CUSUM lines}
}
\description{
displays a straight-forward CUSUM plot
}
\examples{
O <- c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0)
E <- c(0.1, 0.1, 0.1, 0.1, 0.2, 0.2, 0.3, 0.3, 0.4, 0.4, 0.1, 0.1,
       0.1, 0.1, 0.2, 0.2, 0.3, 0.3, 0.4, 0.4)
sor.var <- 1:20
cusumPlot()
}
\keyword{cusum}

