% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.errors.R
\name{dotplot.errors}
\alias{dotplot.errors}
\title{dotplot.errors}
\usage{
\method{dotplot}{errors}(
  x,
  myTheme = simpleTheme(pch = 19, col = 1),
  qlabel = "Estimated rate",
  add.text.to.qlabel = "",
  type.bar = "CI",
  conf.level = 0.95,
  end.length = 0.05,
  reorder.groups = TRUE,
  reordering = "decreasing",
  label.define = list(),
  reference.line = NULL,
  bar.color = 1,
  horizontal = TRUE,
  ...
)
}
\arguments{
\item{x}{dataframe}

\item{myTheme}{default values for pch and col (todo)}

\item{qlabel}{text for axis label}

\item{add.text.to.qlabel}{additional text for axis label}

\item{type.bar}{default is "CI"}

\item{conf.level}{default is .95}

\item{end.length}{default is .05}

\item{reorder.groups}{default is TRUE}

\item{reordering}{order of sorting when reorder.groups is TRUE}

\item{label.define}{(todo)}

\item{reference.line}{value for the reference line. default is null}

\item{bar.color}{default is 1}

\item{horizontal}{layout of the dotplot. default is TRUE}

\item{...}{further arguments passed to or from other methods.}
}
\description{
build a vertical dotplot with error bars
}
\author{
Rollie Parrish
}
