% Generated by roxygen2 (4.0.0): do not edit by hand
\name{edwFetch}
\alias{edwFetch}
\title{edwFetch}
\usage{
edwFetch(resource = "Phloston", schemaName = NULL, tableName = NULL,
  max = 10)
}
\arguments{
\item{resource}{- Alias for the database connection}

\item{schema}{- name of the desired schema}

\item{tableName}{- name of the desired table}

\item{catalog}{- name of the desired database}
}
\description{
read some or all of a table from the EDW into a data frame
}
\examples{
edw_tables("Phloston")
}
\keyword{sql}

