% Generated by roxygen2 (4.0.0): do not edit by hand
\name{edwFetch}
\alias{edwFetch}
\title{edwFetch}
\usage{
edwFetch(resource = "Phloston", schema = NULL, table = NULL, max = 10)
}
\arguments{
  \item{resource}{- Alias for the database connection}

  \item{schema}{- desired schema}

  \item{table}{- desired table}

  \item{max}{= TOP n rows from the query results}
}
\description{
read some or all of a table from the EDW into a data frame
}
\examples{
edwFetch(schemaName="Pharmacy", tableName="MedicationBASE")
}
\keyword{sql}

