% Generated by roxygen2 (4.0.0): do not edit by hand
\name{edwQuery}
\alias{edwQuery}
\title{edwquery}
\usage{
edwQuery(schema, table_name, max = 10, fields = "*", where = "1=1",
  resource = "Phloston")
}
\arguments{
\item{schema}{- name of the desired schema}

\item{table_name}{- name of the desired table}

\item{max}{- used to return the TOP n records - use a large number for all records}

\item{fields}{- single string of field names separated by commas}

\item{where}{- additional where clauses}

\item{resource}{- Alias for the database connection}
}
\description{
queries EDW using windows credentials
}
\examples{
sql_query("")
}
\keyword{sql}

