% Generated by roxygen2 (4.0.0): do not edit by hand
\name{edwQuery}
\alias{edwQuery}
\title{edwquery}
\usage{
edwQuery(schema, table_name, max = 10, fields = "*", where = "1=1",
  order = "", na.rm = TRUE, resource = "Phloston")
}
\arguments{
  \item{schema}{name of the desired schema}

  \item{table_name}{name of the desired table}

  \item{max}{used to return the TOP n records - defaults to
  10. Set to FALSE to return all results}

  \item{fields}{single string of field names separated by
  commas}

  \item{where}{additional WHERE clauses}

  \item{order}{additional ORDER BY clauses}

  \item{na.rm}{remove columns where all values are NA}

  \item{resource}{Alias for the database connection}
}
\value{
Currently, a list is returned with the following elements,
\enumerate{ \item \code{data}: an R \code{data.frame} of
the desired records and columns.  \item \code{fields}: list
of the desired field names.  \item \code{elapsed_seconds}:
the duration of the function.  \item \code{status_message}:
a boolean value indicating if the operation was apparently
successful. }
}
\description{
queries EDW using windows credentials
}
\keyword{sql}

