% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edwSQL.R
\name{edwSQL}
\alias{edwSQL}
\title{edwSQL}
\usage{
edwSQL(sql = "SELECT TOP 10 * FROM Event_Cath", resource = "custom",
  custom = NULL, file = TRUE, DSN = FALSE, uid = NULL, pwd = NULL,
  ...)
}
\arguments{
\item{sql}{- sql query or the path & name of a .sql file}

\item{resource}{- Alias for the database connection

\describe{
  \item{custom}{allows for a connection to any database server using the server parameter}
  \item{EDWDBDev}{EDWDBDev}
  \item{EDWDBProd}{EDWDBProd}
  \item{Trantor}{EDWDBProd}
  \item{Phloston}{EDWDBDev}
  \item{ClinicalAnalytics}{\\\\wn2591\\PremierPRD}
  \item{Apollo}{Apollo_Spokane}
  \item{Eva}{\\\\wn1444 Testing}
}}

\item{custom}{full connection string when using 'custom' as the resource.}

\item{file}{switch indicating if sql parameter is a .sql file or a command. Default is TRUE (requires .sql file)}

\item{DSN}{if TRUE, use a DSN as specified in `resource` parameter instead of a connection string (ie. for Axis PATS: resource = "TSI_32", DSN = TRUE)}

\item{uid}{username for connections if necessary. Default is NULL, which will then use the user's system credentials.}

\item{pwd}{password for connections that don't use the user's system credentials. Default is NULL.}
}
\value{
\item{data}{Query results as a data frame}
\item{fields}{field names}
\item{elapsed_seconds}{elapsed system time to run the query}
\item{status_message}{the number of records & variables in the results}
}
\description{
passes a SQL query to the EDW. Commented lines in the .sql file must be preceded by '--'. Blocked comments are not supported.
}
\keyword{sql}

