% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sql_query.R
\name{sql_query}
\alias{sql_query}
\title{sql_query}
\usage{
sql_query(sql, dsn = "", file = TRUE)
}
\arguments{
\item{sql}{}

\item{dsn}{}

\item{file}{(defaults to TRUE)}
}
\description{
queries SQL Server as configured in system DSN
}
\keyword{sql}

