% Generated by roxygen2 (4.0.0): do not edit by hand
\name{sql_query}
\alias{sql_query}
\title{sql_query}
\usage{
sql_query(sql, dsn = "", file = TRUE)
}
\arguments{
  \item{sql}{}

  \item{dsn}{}

  \item{file}{(defaults to TRUE)}
}
\description{
queries SQL Server as configured in system DSN
}
\examples{
sql_query("select * from table")
}
\keyword{sql}

